% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{changeLog}
\alias{changeLog}
\title{changeLog}
\usage{
changeLog(n=20)
}
\arguments{
\item{n}{numeric, 20 by default, the number of lines will be shown up.}
}
\value{
list, the most recent n lines
}
\description{
Get the most recent n lines
}
\details{
changeLog
}
\note{
Once you have installed cooccurNet, the change-log can also be viewed from the R prompt.
}
\examples{
logs = changeLog(n=20)
}

