% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{getexample_forRCOS}
\alias{getexample_forRCOS}
\title{getexample_forRCOS}
\usage{
getexample_forRCOS()
}
\value{
File paths for all the files available for testing RCOS.
}
\description{
Get the test data for testing RCOS method
}
\details{
getexample_forRCOS
}
\note{
File paths for all the files available for testing RCOS. Currently, there is only one file "HA1protein_humanH3N2", the sequences within which are derived from the database of Influenza Virus Resource.
}
\examples{
dataFile = getexample_forRCOS()
}

