% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{readseq}
\alias{readseq}
\title{readseq}
\usage{
readseq(dataFile="",  dataType="DNA", debug=FALSE)
}
\arguments{
\item{dataFile}{the file name of data}

\item{dataType}{the type of data. It could be 'DNA' (default), 'protein', 'SNP' or 'other'}

\item{debug}{FALSE by default; to indicate whether the debug message will be displayed or not}
}
\value{
a list object containing the original data matrix
}
\description{
Reading a sequence file in fasta format
}
\examples{
data = readseq(dataFile=getexample(dataType="protein"), dataType="protein")
}

