% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{toigraph}
\alias{toigraph}
\title{toigraph}
\usage{
toigraph(networkFile="", networkNames=c())
}
\arguments{
\item{networkFile}{the network file name for the co-occurrence network. It is generated from cooc() or gencooccur().}

\item{networkNames}{a vector of network names}
}
\value{
a list of igraph::graph.data.frame object ordered by the input network names.
}
\description{
transform a network file to the igraph.data.frame type by specifying a network file name and a network name
}
\examples{
#cooccurNetwork  = cooc(dataFile=getexample(dataType="protein"),dataType="protein")
#network_igraph = toigraph(networkFile=cooccurNetwork$networkFile, networkName=c("EPI823725")
}

