% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{gencooccur}
\alias{gencooccur}
\title{gencooccur}
\usage{
gencooccur(data = list(), cooccurFilter = NULL,
  networkFile = "cooccurNetwork", module = FALSE,
  moduleFile = "cooccurNetworkModule", property = FALSE,
  propertyFile = "cooccurNetworkProperty", siteCo = FALSE,
  siteCoFile = "siteCooccurr", sampleTimes = 100, debug = FALSE)
}
\arguments{
\item{data}{a list return from function "pprocess()"}

\item{cooccurFilter}{a number in the range of 0~1. It determines whether two columns are perfect co-occurrence;}

\item{networkFile}{file name with full path for storing the co-occurrence network for each row;}

\item{module}{FALSE by default. If it is set to be TRUE, the modules in each network of the networkFile would be calculated.}

\item{moduleFile}{file name with full path  for storing the modules for co-occurrence network;}

\item{property}{FALSE by default. If it is set to be TRUE, the properties for each network of the networkFile, including the network diameter, connectivity, ConnectionEffcient and so on, would be calculated.}

\item{propertyFile}{character,  file name with full path  for storing the modules for co-occurrence network;}

\item{siteCo}{FALSE by default. If it is set to be TRUE, the extent of co-occurrence between all pairs of columns would be calculated. It is defined as the ratio of rows with perfect co-occurrence.}

\item{siteCoFile}{file name with full path  for storing the extent of co-occurrence between all pairs of columns, and the related p-values. The later are calculated by simulations as follows   firstly, all columns in the data are randomly permutated; then, the pairwise siteCos are calculated. This process would be repeated N times (the value depends on the parameter sampleTimes). For each pair of columns, the rank of the original siteCo in the N siteCos derived from simulations are considered as the p-value for the original siteCo.}

\item{sampleTimes}{a integer of permutations in the simulation when calculating the p-values.}

\item{debug}{FALSE by default; to indicate whether the debug message will be displayed or not}
}
\value{
a list and all the output file paths are attributed in it. \cr  The attribute "networkFile" stores the co-occurrence network for each row; \cr The attribute "moduleFile" is optional. When the module is set to be TRUE, it would be output. It stores the modules for co-occurrence network; \cr The attribute "propertyFile" is optional. When the property is set to be TRUE, it would be output. It stores the properties for co-occurrence network; \cr The attribute "siteCoFile" is optional. When the property is set to be TRUE, it would be output. It stores all the pairwise siteCos between columns.
}
\description{
Construct the co-occurrence network
}
\examples{
data = readseq(dataFile=getexample(dataType="protein"),dataType="protein")
data_process = pprocess(data=data)
#cooccurNetwork  = gencooccur(data=data_process)
}
\references{
Du, X., Wang, Z., Wu, A., Song, L., Cao, Y., Hang, H., & Jiang, T. (2008). Networks of genomic co-occurrence capture characteristics of human influenza A (H3N2) evolution. Genome research, 18(1), 178-187. doi:10.1101/gr.6969007
}

