% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{pprocess}
\alias{pprocess}
\title{pprocess}
\usage{
pprocess(data=list(),  conservativeFilter=0.95, debug=FALSE)
}
\arguments{
\item{data}{a list  return from function "readseq()"}

\item{conservativeFilter}{0.95 by default. a number in the range of 0~1. The column with conservative score greater than it would be filtered in the later analyses;}

\item{debug}{FALSE by default; to indicate whether the debug message will be displayed or not}
}
\value{
a list object containing the original data matrix and frequency matrix
}
\description{
Filter the conservative columns (defined as the conservative score greater than the "conservativeFilter")
}
\examples{
data = readseq(dataFile=getexample(dataType="protein"),dataType="protein")
data_process = pprocess(data=data)
}
\references{
Du, X., Wang, Z., Wu, A., Song, L., Cao, Y., Hang, H., & Jiang, T. (2008). Networks of genomic co-occurrence capture characteristics of human influenza A (H3N2) evolution. Genome research, 18(1), 178-187. doi:10.1101/gr.6969007
}

