% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{readseq}
\alias{readseq}
\title{readseq}
\usage{
readseq(dataFile="", dataType="protein", debug=FALSE)
}
\arguments{
\item{dataFile}{character, a FASTA data file name with full path.}

\item{dataType}{character, 'protein' by default, the type of data will be processed. It could be 'DNA', 'RNA', 'protein', 'SNP' or 'other'.}

\item{debug}{logic, FALSE by default, to indicate whether the debug message will be displayed or not.}
}
\value{
list,contains the original data matrix and some other information.
}
\description{
Reading a sequence file in fasta format
}
\examples{
data = readseq(dataFile=getexample(dataType="protein"), dataType="protein")
}

